requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.2
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.8
Pygments==2.15.1
python-dotenv==1.0.0
qdrant-client==1.2.0
regex==2023.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tiktoken==0.4.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.16
wcwidth==0.2.6
yarl==1.9.2
